Remove Phone to Watch Sync It does not work on any reasonably modern device
