+++ dndsync/common/src/main/kotlin/com/boswelja/smartwatchextensions/dndsync/Paths.kt
@@ -9,8 +9,3 @@ const val DnDStatusPath = "/dnd_status"
  * A path for devices to declare they can send DnD state
  */
 const val SendDnDCapability = "SEND_DND"
-
-/**
- * A path for devices to declare they can receive DnD state
- */
-const val ReceiveDnDCapability = "RECEIVE_DND"
