+++ dndsync/mobile/src/main/AndroidManifest.xml
@@ -9,9 +9,5 @@
             </intent-filter>
         </receiver>
 
-        <service
-            android:name=".LocalDnDCollectorService"
-            android:exported="false" />
-
     </application>
 </manifest>
