+++ dndsync/wear/src/main/AndroidManifest.xml
@@ -1,14 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
 
-        <receiver
-            android:name=".DnDStatusReceiver"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.boswelja.watchconnection.common.message.MessageReceived" />
-            </intent-filter>
-        </receiver>
-
         <service
             android:name=".LocalDnDAndTheaterCollectorService"
             android:exported="false" />
