--- dndsync/wear/src/main/kotlin/com/boswelja/smartwatchextensions/dndsync/DnDStatusReceiver.kt
@@ -1,6 +1,0 @@
-package com.boswelja.smartwatchextensions.dndsync
-
-/**
- * An implementation of [BaseDnDStatusReceiver] for wearables.
- */
-class DnDStatusReceiver : BaseDnDStatusReceiver()
