+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/bootorupdate/Utils.kt
@@ -1,11 +1,7 @@
 package com.boswelja.smartwatchextensions.bootorupdate
 
 import android.content.Context
-import android.content.Intent
-import androidx.core.content.ContextCompat
 import com.boswelja.smartwatchextensions.core.settings.WatchSettingsRepository
-import com.boswelja.smartwatchextensions.dndsync.DnDSyncSettingKeys.DND_SYNC_TO_WATCH_KEY
-import com.boswelja.smartwatchextensions.dndsync.LocalDnDCollectorService
 import com.boswelja.smartwatchextensions.proximity.SeparationObserverService
 import com.boswelja.smartwatchextensions.proximity.common.ProximitySettingKeys.WATCH_SEPARATION_NOTI_KEY
 import kotlinx.coroutines.flow.first
@@ -21,22 +17,7 @@ internal const val NotiId = 69102
 internal suspend fun Context.restartServices(
     settingsRepository: WatchSettingsRepository
 ) {
-    tryStartInterruptFilterSync(settingsRepository)
     tryStartSeparationObserver(settingsRepository)
-}
-
-/** Try to start Do not Disturb change listener service if needed. */
-private suspend fun Context.tryStartInterruptFilterSync(repository: WatchSettingsRepository) {
-    val dndSyncToWatchEnabled = repository
-        .getIdsWithBooleanSet(DND_SYNC_TO_WATCH_KEY, true)
-        .map { it.isNotEmpty() }
-        .first()
-    if (dndSyncToWatchEnabled) {
-        ContextCompat.startForegroundService(
-            applicationContext,
-            Intent(applicationContext, LocalDnDCollectorService::class.java)
-        )
-    }
 }
 
 private suspend fun Context.tryStartSeparationObserver(repository: WatchSettingsRepository) {
